+++ NEWS.md
@@ -3,7 +3,9 @@
 * standardised file names
 * much improved documentation
 * more coherent `makeCPO*` functions
-* %<>>% and %<<<% operators for compose / attach + assignment 
+* %<>>% and %<<<% operators for compose / attach + assignment
+* More coherent handling of cababilities
+* New getters: getCPOTrainedCapability, getOriginalCPO, getOriginalCPOConstructor
 
 # mlrCPO 0.1
 * Initial release on GitHub, retrafoless and target operating CPOs don't work yet